+++ src/tools/ft_memset.c
@@ -12,9 +12,9 @@
 
 #include "so_long.h"
 
-void	*ft_memset(void *s, int c, size_t n)
+void	*ft_memset(void *s, int c, int n)
 {
-	size_t	i;
+	int	i;
 	char	*ptr;
 
 	ptr = (char *)s;
